+++ crates/mimic_build/src/actor/mod.rs
@@ -1,7 +1,7 @@
 pub mod db;
 pub mod fixtures;
+pub mod metrics;
 pub mod query;
-pub mod stats;
 
 use mimic_schema::{
     build::get_schema,
@@ -52,7 +52,7 @@ impl ActorBuilder {
 
         // shared between all canisters
         tokens.extend(db::generate(&self));
-        tokens.extend(stats::generate(&self));
+        tokens.extend(metrics::generate(&self));
         tokens.extend(fixtures::generate(&self));
         tokens.extend(query::generate(&self));