+++ crates/mimic_schema/src/node/item.rs
@@ -4,7 +4,7 @@ use crate::{
     types::Primitive,
     visit::Visitor,
 };
-use mimic_common::error::ErrorTree;
+use mimic_common::{err, error::ErrorTree};
 use serde::Serialize;
 use std::ops::Not;
 
@@ -48,7 +48,7 @@ impl ValidateNode for Item {
             ItemTarget::Is(path) => {
                 // cannot be an entity
                 if schema.check_node_as::<Entity>(path).is_ok() {
-                    errs.add("a non-relation Item cannot reference an Entity");
+                    err!(errs, "a non-relation Item cannot reference an Entity");
                 }
 
                 // todo
@@ -56,12 +56,13 @@ impl ValidateNode for Item {
                     match node.get_type() {
                         Some(tnode) => {
                             if !self.todo && tnode.ty().todo {
-                                errs.add(format!(
+                                err!(
+                                    errs,
                                     "you must specify todo if targeting a todo flagged item ({path})",
-                                ));
+                                );
                             }
                         }
-                        None => errs.add("node is not a valid type"),
+                        None => err!(errs, "node is not a valid type"),
                     }
                 }
             }
@@ -72,7 +73,7 @@ impl ValidateNode for Item {
         // relation
         if let Some(relation) = &self.relation {
             if self.indirect {
-                errs.add("relations cannot be set to indirect");
+                err!(errs, "relations cannot be set to indirect");
             }
 
             // Step 1: Ensure the relation path exists and is an Entity
@@ -84,14 +85,16 @@ impl ValidateNode for Item {
 
                     // Step 3: Compare to self.target()
                     if &self.target != relation_target {
-                        errs.add(format!(
+                        err!(
+                            errs,
                             "relation target type mismatch: expected {:?}, found {:?}",
-                            relation_target, self.target
-                        ));
+                            relation_target,
+                            self.target
+                        );
                     }
                 }
                 None => {
-                    errs.add(format!("relation entity '{relation}' not found"));
+                    err!(errs, "relation entity '{relation}' not found");
                 }
             }
         }
@@ -100,7 +103,7 @@ impl ValidateNode for Item {
         if let Some(selector) = &self.selector
             && schema.get_node_as::<Selector>(selector).is_none()
         {
-            errs.add(format!("selector path '{selector}' not found"));
+            err!(errs, "selector path '{selector}' not found");
         }
 
         errs.result()